added unit Tests
